requirements.txt
qdrant-client
fastembed
kagglehub
streamlit
python-dotenv
Pillow